refactor: exhibition API 타입 분리 (exhibition.types.ts)
