+++ apps/client/lib/api/exhibition.ts
@@ -1,26 +1,13 @@
 import { apiClient } from "api";
+import type {
+	BannerItem,
+	ExhibitionDetail,
+	ExhibitionHost,
+	ExhibitionItem,
+	HostSns,
+} from "./exhibition.types";
 
-export interface BannerItem {
-	id: number;
-	imageUrl: string;
-	orderIndex: number;
-	linkUrl: string;
-}
-
-export interface ExhibitionItem {
-	id: string;
-	slug: string;
-	title: string;
-	univName: string;
-	deptName: string;
-	exhibitionType?: string | null;
-	imageUrl: string | null;
-	logoImg: string;
-	startDate: string | null;
-	endDate: string | null;
-	address: string | null;
-	dday: number | null;
-}
+export type * from "./exhibition.types";
 
 export async function getMainExhibitions(): Promise<ExhibitionItem[]> {
 	try {
@@ -61,42 +48,6 @@ export async function resolveExhibitionSlug(slug: string): Promise<{ uuid: strin
  * GET exhibitions/{exhibitionId}/host
  * GET exhibitions/{exhibitionId}/host/sns
  ************************/
-
-export interface ExhibitionLocation {
-	address: string; // 기본 주소
-	latitude: string;
-	longitude: string;
-	detail_location: string | null; // 상세 주소
-	location_description: string | null; // 찾아오는 길 설명
-}
-
-export interface ExhibitionHost {
-	hostId: string;
-	hostName: string;
-	hostImageUrl: string;
-	description: string;
-	email: string;
-}
-export interface HostSns {
-	snsId: string;
-	platformName: string;
-	url: string;
-}
-
-export interface ExhibitionDetail {
-	exhibitionId: string;
-	univName: string;
-	deptName: string;
-	exhibitionType?: string | null;
-	title: string;
-	exhibitionImg: string;
-	startDate: string;
-	endDate: string;
-	dateInfo: string;
-	description: string;
-	location: ExhibitionLocation;
-	isPublic: boolean;
-}
 
 export async function getExhibitionDetail(exhibitionId: string): Promise<ExhibitionDetail | null> {
 	try {
